test-requirements.txt
# needed for doc build
Jinja2>=2.7.3
PyYAML==3.11
sphinx>=1.1.2,!=1.2.0,<1.3
oslosphinx>=2.2.0  # Apache-2.0
